requirements.txt
# ── ИИ-ядра ───────────────────────────────────────────────
google-generativeai>=0.7.0

# ── Голос (TTS + STT) ─────────────────────────────────────
pyttsx3>=2.90
PyAudio>=0.2.14
SpeechRecognition>=3.10.0

# ── Интерфейс Desktop ─────────────────────────────────────
customtkinter>=5.2.0

# ── Telegram ──────────────────────────────────────────────
python-telegram-bot>=21.0

# ── Сеть и парсинг ────────────────────────────────────────
requests>=2.31.0
beautifulsoup4>=4.12.0

# ── Системный мониторинг ──────────────────────────────────
psutil>=5.9.0

# ── IoT / COM-порты ───────────────────────────────────────
pyserial>=3.5

# ── Безопасность и шифрование ─────────────────────────────
cryptography>=42.0.0
python-dotenv>=1.0.0

# ── Android (только для мобильной сборки) ─────────────────
# kivy>=2.3.0
# plyer>=2.1.0

# ── Сборка .exe ───────────────────────────────────────────
# pyinstaller>=6.0.0
